am335x: switch to /usr/bin/bb_dnsmasq_config.sh

+++ boot/am335x_evm.sh
@@ -724,43 +724,45 @@ if [ ! "x${USB_NETWORK_DISABLED}" = "xyes" ]; then
 					echo "${log} dnsmasq: setting up for usb0/usb1"
 					disable_connman_dnsproxy
 
-					wfile="/etc/dnsmasq.d/SoftAp0"
-					echo "interface=usb0" > ${wfile}
+					if [ -f /usr/bin/bb_dnsmasq_config.sh ] ; then
+						/usr/bin/bb_dnsmasq_config.sh || true
+					else
+						wfile="/etc/dnsmasq.d/SoftAp0"
+						echo "interface=usb0" > ${wfile}
 
-					if [ "x${USB1_ENABLE}" = "xenable" ] ; then
-						echo "interface=usb1" >> ${wfile}
+						if [ "x${USB1_ENABLE}" = "xenable" ] ; then
+							echo "interface=usb1" >> ${wfile}
+						fi
+
+						echo "port=53" >> ${wfile}
+						echo "dhcp-authoritative" >> ${wfile}
+						echo "domain-needed" >> ${wfile}
+						echo "bogus-priv" >> ${wfile}
+						echo "expand-hosts" >> ${wfile}
+						echo "cache-size=2048" >> ${wfile}
+						echo "dhcp-range=usb0,${USB0_SUBNET}.1,${USB0_SUBNET}.1,2m" >> ${wfile}
+
+						if [ "x${USB1_ENABLE}" = "xenable" ] ; then
+							echo "dhcp-range=usb1,${USB1_SUBNET}.1,${USB1_SUBNET}.1,2m" >> ${wfile}
+						fi
+
+						echo "listen-address=127.0.0.1" >> ${wfile}
+						echo "listen-address=${USB0_SUBNET}.2" >> ${wfile}
+
+						if [ "x${USB1_ENABLE}" = "xenable" ] ; then
+							echo "listen-address=${USB1_SUBNET}.2" >> ${wfile}
+						fi
+
+						echo "dhcp-option=usb0,3" >> ${wfile}
+						echo "dhcp-option=usb0,6" >> ${wfile}
+
+						if [ "x${USB1_ENABLE}" = "xenable" ] ; then
+							echo "dhcp-option=usb1,3" >> ${wfile}
+							echo "dhcp-option=usb1,6" >> ${wfile}
+						fi
+
+						echo "dhcp-leasefile=/var/run/dnsmasq.leases" >> ${wfile}
 					fi
-
-					echo "port=53" >> ${wfile}
-					echo "dhcp-authoritative" >> ${wfile}
-					echo "domain-needed" >> ${wfile}
-					echo "bogus-priv" >> ${wfile}
-					echo "expand-hosts" >> ${wfile}
-					echo "cache-size=2048" >> ${wfile}
-					echo "dhcp-range=usb0,${USB0_SUBNET}.1,${USB0_SUBNET}.1,2m" >> ${wfile}
-
-					if [ "x${USB1_ENABLE}" = "xenable" ] ; then
-						echo "dhcp-range=usb1,${USB1_SUBNET}.1,${USB1_SUBNET}.1,2m" >> ${wfile}
-					fi
-
-					echo "listen-address=127.0.0.1" >> ${wfile}
-					echo "listen-address=${USB0_SUBNET}.2" >> ${wfile}
-
-					if [ "x${USB1_ENABLE}" = "xenable" ] ; then
-						echo "listen-address=${USB1_SUBNET}.2" >> ${wfile}
-					fi
-
-					echo "dhcp-option=usb0,3" >> ${wfile}
-					echo "dhcp-option=usb0,6" >> ${wfile}
-
-					if [ "x${USB1_ENABLE}" = "xenable" ] ; then
-						echo "dhcp-option=usb1,3" >> ${wfile}
-						echo "dhcp-option=usb1,6" >> ${wfile}
-					fi
-
-		#FIXME: why was this added, without connman every ip get's 172.1.8.1????
-		#			echo "address=/#/172.1.8.1" >> ${wfile}
-					echo "dhcp-leasefile=/var/run/dnsmasq.leases" >> ${wfile}
 
 					systemctl restart dnsmasq || true
 				else
